fix: widen nutrition meal_type and unit_code to integer
Les codes Samsung sont à 6 chiffres — meal_type 100001 à 100006, unit_code
120001 à 120005 — alors que smallint plafonne à 32 767. Les 11 855 entrées
alimentaires étaient donc impossibles à ingérer.

La correction est portée dans la migration initiale plutôt que dans une
migration de rattrapage : rien n'est déployé, et un schéma initial correct
vaut mieux qu'un schéma faux suivi d'un rustine pour un état qui n'a jamais
existé ailleurs.

Ajoute aussi scripts/migrate_legacy.py, l'utilitaire de migration one-shot.

+++ backend/app/models/nutrition.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import BigInteger, DateTime, Double, SmallInteger, Text
+from sqlalchemy import BigInteger, DateTime, Double, Integer, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base, TimestampMixin
@@ -21,8 +21,12 @@ class NutritionEntry(Base, TimestampMixin):
         DateTime(timezone=True), nullable=False, index=True
     )
 
-    meal_type: Mapped[int | None] = mapped_column(SmallInteger, index=True)
+    # Codes Samsung à 6 chiffres (ex. 100001 pour meal_type, 120001 pour
+    # unit_code) : un SmallInteger (max 32 767) déborde sur les vraies
+    # données, cf. sport_type qui souffre du même problème et est déjà en
+    # Integer.
+    meal_type: Mapped[int | None] = mapped_column(Integer, index=True)
     food_name: Mapped[str] = mapped_column(Text, nullable=False, default="")
     amount: Mapped[float | None] = mapped_column(Double)
-    unit_code: Mapped[int | None] = mapped_column(SmallInteger)
+    unit_code: Mapped[int | None] = mapped_column(Integer)
     calories: Mapped[float | None] = mapped_column(Double)
